refactor: change package structure of nextstep.jdbc
