+++ nextstep-jdbc/src/main/java/nextstep/jdbc/exception/JdbcTemplateSqlException.java
@@ -1,4 +1,4 @@
-package nextstep.jdbc;
+package nextstep.jdbc.exception;
 
 public class JdbcTemplateSqlException extends RuntimeException {
     public JdbcTemplateSqlException(Throwable cause) {
