+++ nextstep-jdbc/src/main/java/nextstep/jdbc/mapper/ObjectMapper.java
@@ -1,4 +1,4 @@
-package nextstep.jdbc;
+package nextstep.jdbc.mapper;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
