+++ nextstep-jdbc/src/main/java/nextstep/jdbc/query/SqlMapper.java
@@ -1,4 +1,4 @@
-package nextstep.jdbc;
+package nextstep.jdbc.query;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
